homework/hw3/pyproject.toml
[project]
name = "hw3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-core>=2.29.0",
    "google-cloud-storage>=3.9.0",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
